Improve the AIDeployment Manager
