+++ src/Primitives/CrestApps.Core.AI/Indexing/AIDocumentSearchIndexProfileHandler.cs
@@ -10,7 +10,11 @@ namespace CrestApps.Core.AI.Indexing;
 
 public sealed class AIDocumentSearchIndexProfileHandler : EmbeddingSearchIndexProfileHandlerBase
 {
-    public AIDocumentSearchIndexProfileHandler(ICatalog<AIDeployment> deploymentCatalog, IAIClientFactory aiClientFactory, ILogger<AIDocumentSearchIndexProfileHandler> logger) : base(IndexProfileTypes.AIDocuments, deploymentCatalog, aiClientFactory, logger)
+    public AIDocumentSearchIndexProfileHandler(
+        ICatalog<AIDeployment> deploymentCatalog,
+        IAIClientFactory aiClientFactory,
+        ILogger<AIDocumentSearchIndexProfileHandler> logger)
+        : base(IndexProfileTypes.AIDocuments, deploymentCatalog, aiClientFactory, logger)
     {
     }
