+++ src/Primitives/CrestApps.Core.AI/Indexing/AIMemorySearchIndexProfileHandler.cs
@@ -16,7 +16,11 @@ public sealed class AIMemorySearchIndexProfileHandler : EmbeddingSearchIndexProf
     private const string _contentFieldName = "content";
     private const string _updatedUtcFieldName = "updatedUtc";
     private const string _embeddingFieldName = "embedding";
-    public AIMemorySearchIndexProfileHandler(ICatalog<AIDeployment> deploymentCatalog, IAIClientFactory aiClientFactory, ILogger<AIMemorySearchIndexProfileHandler> logger) : base(IndexProfileTypes.AIMemory, deploymentCatalog, aiClientFactory, logger)
+    public AIMemorySearchIndexProfileHandler(
+        ICatalog<AIDeployment> deploymentCatalog,
+        IAIClientFactory aiClientFactory,
+        ILogger<AIMemorySearchIndexProfileHandler> logger)
+        : base(IndexProfileTypes.AIMemory, deploymentCatalog, aiClientFactory, logger)
     {
     }
