+++ src/Primitives/CrestApps.Core.AI/Indexing/EmbeddingSearchIndexProfileHandlerBase.cs
@@ -15,7 +15,11 @@ public abstract class EmbeddingSearchIndexProfileHandlerBase : IndexProfileHandl
     private readonly ICatalog<AIDeployment> _deploymentCatalog;
     private readonly IAIClientFactory _aiClientFactory;
     private readonly ILogger _logger;
-    protected EmbeddingSearchIndexProfileHandlerBase(string type, ICatalog<AIDeployment> deploymentCatalog, IAIClientFactory aiClientFactory, ILogger logger)
+    protected EmbeddingSearchIndexProfileHandlerBase(
+        string type,
+        ICatalog<AIDeployment> deploymentCatalog,
+        IAIClientFactory aiClientFactory,
+        ILogger logger)
     {
         _type = type;
         _deploymentCatalog = deploymentCatalog;
